+++ src/components/pagelayout/PageLayout.tsx
@@ -1,4 +1,4 @@
-import Sidebar from "../sidebar/Sidebar";
+import Sidebar from "./Sidebar";
 
 interface PageLayoutProps {
   title: string;
@@ -6,15 +6,15 @@ interface PageLayoutProps {
   children: React.ReactNode;
 }
 
-const PageLayout: React.FC<PageLayoutProps> = ({ title,  children }) => {
+const PageLayout: React.FC<PageLayoutProps> = ({ title, children }) => {
   return (
     <div className="flex min-h-screen font-sans">
       {/* ✅ 공통 사이드바 */}
-      <Sidebar   />
+      <Sidebar />
 
       {/* ✅ 메인 콘텐츠 영역 */}
-      <main className="flex-1 ml-34 p-8 bg-white">
-        <div className="max-w-7xl mx-auto">
+      <main className="flex-1  p-10 bg-white">
+        <div className="max-w-7xl mx-3 ">
           {/* ✅ 공통 제목 바 */}
           <div className="flex flex-col mb-6">
             <h1 className="text-2xl font-bold">{title}</h1>
