+++ src/components/pagelayout/Sidebar.tsx
@@ -1,43 +1,7 @@
 import { FC } from "react";
 import { Link, useLocation } from "react-router-dom";
 import { FaFileAlt } from "react-icons/fa"; // 아이콘 추가
-
-const menuData = [
-  {
-    name: "소개",
-    path: "/introduce",
-    links: [
-      { to: "/introduce/project", label: "프로젝트 소개" },
-      { to: "/introduce/team", label: "팀소개" },
-    ],
-  },
-  {
-    name: "AI",
-    path: "/upload",
-    links: [
-      { to: "/upload/form", label: "이미지 업로드" },
-      { to: "/upload/list", label: "분석결과 현황" },
-    ],
-  },
-  {
-    name: "알림소식",
-    path: "/board",
-    links: [
-      { to: "/board/notice", label: "공지사항" },
-      { to: "/board/list", label: "게시판" },
-    ],
-  },
-  {
-    name: "고객지원",
-    path: "/support",
-    links: [
-      { to: "/pract", label: "연습" },
-      { to: "/practice", label: "연습2" },
-      { to: "/prac", label: "연습3" },
-      { to: "/supportmenu", label: "사이드바" },
-    ],
-  },
-];
+import { menuData } from "./MenuData";
 
 const Sidebar: FC = () => {
   const location = useLocation();
@@ -64,7 +28,7 @@ const Sidebar: FC = () => {
                 className={`flex items-center px-4 py-3 text-sm font-medium rounded-lg shadow-md my-3 border-2 transition ${
                   location.pathname === to 
                     ? "border-blue-600 text-blue-600 bg-white"  // 선택된 메뉴: 파란색 테두리 + 파란색 글씨
-                    : "border-transparent text-gray-700 hover:border-blue-400 hover:text-blue-600"
+                    : "border-gray-300 text-gray-700 hover:border-blue-400 hover:text-blue-600"
                 } transition`}
               >
                 <span className="ml-3">{label}</span>
