+++ src/pages/upload/UploadForm.tsx
@@ -2,10 +2,10 @@ import PageLayout from "../../components/pagelayout/PageLayout";
 
 const UploadForm = () => {
   return (
-    <PageLayout title="분석결과 현황" activeItem="분석결과 현황">
+    <PageLayout title="이미지 업로드" activeItem="이미지 업로드">
       <div>
         {/* ✅ 여기에 파일 업로드 관련 UI를 추가 */}
-        <p>분석결과 현황 내용</p>
+        <p>이미지 업로드 내용</p>
       </div>
     </PageLayout>
   );
