+++ src/pages/upload/UploadList.tsx
@@ -2,10 +2,10 @@ import PageLayout from "../../components/pagelayout/PageLayout";
 
 const UploadList = () => {
   return (
-    <PageLayout title="파일 업로드" activeItem="파일 업로드">
+    <PageLayout title="분석결과 현황" activeItem="분석결과 현황">
       <div>
         {/* ✅ 여기에 파일 업로드 관련 UI를 추가 */}
-        <p>파일 업로드 페이지 내용</p>
+        <p>분석결과 현황 내용</p>
       </div>
     </PageLayout>
   );
